Reorganised ability files and some sprite files
