+++ include/entity/player/sprites/Facing.h
