--- src/entity/player/Ability.cpp
@@ -1,8 +1,0 @@
-//
-// Created by Agamjeet Singh on 17/08/25.
-//
-
-#include "../../../include/entity/player/Ability.h"
-#include "../../../include/entity/player/Player.h"
-
-Ability::Ability(Player &player): player(player) {}
