+++ src/entity/player/ability/Ability.cpp
@@ -1,0 +1,8 @@
+//
+// Created by Agamjeet Singh on 17/08/25.
+//
+
+#include "../../../../include/entity/player/ability/Ability.h"
+#include "../../../../include/entity/player/Player.h"
+
+Ability::Ability(Player &player): player(player) {}
